move constants for default ROS project layout
