+++ rasa/cli/scaffold.py
@@ -6,8 +6,12 @@ import rasa.train
 from rasa.cli.shell import shell
 from rasa.cli.utils import create_output_path
 from rasa.shared.utils.cli import print_success, print_error_and_exit
-from rasa.constants import DEFAULT_CONFIG_PATH, DEFAULT_DATA_PATH, DEFAULT_DOMAIN_PATH
-from rasa.shared.constants import DOCS_BASE_URL
+from rasa.shared.constants import (
+    DOCS_BASE_URL,
+    DEFAULT_CONFIG_PATH,
+    DEFAULT_DOMAIN_PATH,
+    DEFAULT_DATA_PATH,
+)
 
 
 # noinspection PyProtectedMember
