+++ rasa/cli/test.py
@@ -6,13 +6,14 @@ from typing import List
 import rasa.shared.data
 import rasa.shared.utils.io
 from rasa.cli.arguments import test as arguments
-from rasa.constants import (
+from rasa.shared.constants import (
+    CONFIG_SCHEMA_FILE,
+    DEFAULT_E2E_TESTS_PATH,
     DEFAULT_CONFIG_PATH,
-    DEFAULT_DATA_PATH,
     DEFAULT_MODELS_PATH,
+    DEFAULT_DATA_PATH,
     DEFAULT_RESULTS_PATH,
 )
-from rasa.shared.constants import CONFIG_SCHEMA_FILE, DEFAULT_E2E_TESTS_PATH
 import rasa.shared.utils.validation as validation_utils
 import rasa.cli.utils as cli_utils
 import rasa.utils.io as io_utils
