+++ rasa/cli/train.py
@@ -5,11 +5,13 @@ import rasa.cli.arguments.train as train_arguments
 
 from rasa.cli.utils import get_validated_path, missing_config_keys
 from rasa.shared.utils.cli import print_error
-from rasa.constants import DEFAULT_CONFIG_PATH, DEFAULT_DATA_PATH, DEFAULT_DOMAIN_PATH
 from rasa.shared.constants import (
     CONFIG_MANDATORY_KEYS_CORE,
     CONFIG_MANDATORY_KEYS_NLU,
     CONFIG_MANDATORY_KEYS,
+    DEFAULT_CONFIG_PATH,
+    DEFAULT_DOMAIN_PATH,
+    DEFAULT_DATA_PATH,
 )
