+++ rasa/constants.py
@@ -1,17 +1,5 @@
 import os
 
-DEFAULT_ENDPOINTS_PATH = "endpoints.yml"
-DEFAULT_CREDENTIALS_PATH = "credentials.yml"
-DEFAULT_CONFIG_PATH = "config.yml"
-DEFAULT_DOMAIN_PATH = "domain.yml"
-DEFAULT_ACTIONS_PATH = "actions"
-DEFAULT_MODELS_PATH = "models"
-DEFAULT_CONVERTED_DATA_PATH = "converted_data"
-DEFAULT_DATA_PATH = "data"
-DEFAULT_RESULTS_PATH = "results"
-DEFAULT_NLU_RESULTS_PATH = "nlu_comparison_results"
-DEFAULT_CORE_SUBDIRECTORY_NAME = "core"
-DEFAULT_NLU_SUBDIRECTORY_NAME = "nlu"
 DEFAULT_REQUEST_TIMEOUT = 60 * 5  # 5 minutes
 DEFAULT_RESPONSE_TIMEOUT = 60 * 60  # 1 hour
