+++ rasa/model.py
@@ -11,13 +11,15 @@ import rasa.shared.utils.io
 import rasa.utils.io
 from rasa.cli.utils import create_output_path
 from rasa.shared.utils.cli import print_success
-from rasa.constants import (
-    DEFAULT_MODELS_PATH,
+from rasa.shared.constants import (
+    CONFIG_KEYS_CORE,
+    CONFIG_KEYS_NLU,
+    CONFIG_KEYS,
     DEFAULT_DOMAIN_PATH,
+    DEFAULT_MODELS_PATH,
     DEFAULT_CORE_SUBDIRECTORY_NAME,
     DEFAULT_NLU_SUBDIRECTORY_NAME,
 )
-from rasa.shared.constants import CONFIG_KEYS_CORE, CONFIG_KEYS_NLU, CONFIG_KEYS
 
 from rasa.core.utils import get_dict_hash
 from rasa.exceptions import ModelNotFound
