+++ rasa/shared/constants.py
@@ -54,3 +54,17 @@ CONFIG_KEYS = CONFIG_KEYS_CORE + CONFIG_KEYS_NLU
 CONFIG_MANDATORY_KEYS_CORE = []
 CONFIG_MANDATORY_KEYS_NLU = ["language"]
 CONFIG_MANDATORY_KEYS = CONFIG_MANDATORY_KEYS_CORE + CONFIG_MANDATORY_KEYS_NLU
+
+# Constants for default Rasa Open Source project layout
+DEFAULT_ENDPOINTS_PATH = "endpoints.yml"
+DEFAULT_CREDENTIALS_PATH = "credentials.yml"
+DEFAULT_CONFIG_PATH = "config.yml"
+DEFAULT_DOMAIN_PATH = "domain.yml"
+DEFAULT_ACTIONS_PATH = "actions"
+DEFAULT_MODELS_PATH = "models"
+DEFAULT_CONVERTED_DATA_PATH = "converted_data"
+DEFAULT_DATA_PATH = "data"
+DEFAULT_RESULTS_PATH = "results"
+DEFAULT_NLU_RESULTS_PATH = "nlu_comparison_results"
+DEFAULT_CORE_SUBDIRECTORY_NAME = "core"
+DEFAULT_NLU_SUBDIRECTORY_NAME = "nlu"
