+++ tests/core/test_model.py
@@ -11,10 +11,10 @@ import pytest
 
 from rasa.shared.importers.importer import TrainingDataImporter
 from rasa.shared.importers.rasa import RasaFileImporter
-from rasa.constants import (
+from rasa.shared.constants import (
     DEFAULT_CONFIG_PATH,
-    DEFAULT_DATA_PATH,
     DEFAULT_DOMAIN_PATH,
+    DEFAULT_DATA_PATH,
     DEFAULT_CORE_SUBDIRECTORY_NAME,
 )
 from rasa.shared.core.domain import KEY_RESPONSES
