+++ tests/shared/importers/test_rasa.py
@@ -2,7 +2,11 @@ from pathlib import Path
 from typing import Text
 import os
 
-from rasa.constants import DEFAULT_CONFIG_PATH, DEFAULT_DOMAIN_PATH, DEFAULT_DATA_PATH
+from rasa.shared.constants import (
+    DEFAULT_CONFIG_PATH,
+    DEFAULT_DOMAIN_PATH,
+    DEFAULT_DATA_PATH,
+)
 from rasa.shared.core.constants import DEFAULT_INTENTS
 from rasa.shared.core.domain import Domain
 from rasa.shared.importers.importer import TrainingDataImporter
